feat: move theme classe in core module
